batch: Extract ownership reference records
The ownership package now has boundary tests that can recognize a dedicated
module for baseline reference metadata while the core ownership model still
owns the record.

Baseline reference serialization sits beside batch ownership aggregation,
deletion loading, and replacement-unit normalization. That placement keeps
source-coordinate metadata tied to the largest ownership module and blocks
moving claim records without another dependency back to ownership.

This commit moves BaselineReference into batch.ownership_references and
updates batch callers to import it from that module with a private alias
inside ownership.py.

Subsequent commits will update tests and tighten the architecture guards so
the old ownership module cannot become a reference import surface again.
