+++ src/git_stage_batch/batch/hunk_ownership_translation.py
@@ -12,7 +12,6 @@ from .absence_content import (
 )
 from .ownership import (
     AbsenceClaim,
-    BaselineReference,
     BatchOwnership,
     ReplacementUnit,
     ReplacementUnitOrigin,
@@ -25,6 +24,7 @@ from .ownership_line_entries import (
     old_line_content_by_number as _old_line_content_by_number,
     replacement_unit_origin_for_line_run as _replacement_unit_origin_for_line_run,
 )
+from .ownership_references import BaselineReference
 from .ownership_replacement_units import normalize_replacement_units
 from .replacement_line_runs import ReplacementLineRun as _ReplacementLineRun
