+++ src/git_stage_batch/batch/ownership_merging.py
@@ -8,11 +8,11 @@ from hashlib import sha256
 from ..core.buffer import buffer_byte_chunks
 from .ownership import (
     AbsenceClaim,
-    BaselineReference,
     BatchOwnership,
     ReplacementUnit,
 )
 from .ownership_claims import presence_claims_from_source_lines
+from .ownership_references import BaselineReference
 from .ownership_replacement_units import normalize_replacement_units
